WinRT library now POSTs exception correctly
Both class libraries now coexist, are built correctly (using json.net)
and the output binaries can be used in both wpf and winrt projects. The
MessageDetails that contain the RaygunRequestMessage are still used and
sent with the wpf/winforms/asp.net library, but the WinRT lib doesn't
send anything for this.

I think we might need another object to convey information when the
exception is sent from a desktop project, as opposed to the web stuff in
RaygunRequestMessage?