Added a just-before-sending filter for RaygunMessages.
I don't like that it's a static Func, but I couldn't think of any better options that would be picked up by the RaygunHttpModule.

We need this for security reasons, to add a second layer of checks over the standard IgnoreFormNames. The advantages of providing custom code (instead of strings in config) for this are:
* We can cancel an entire message
* We can log an error if the app tries to send any form variable that matches a particular regex (such as a credit card number)
* We can mutate the message before sending - for example anonymize a section of a single value