Build the message as part of the work item
Currently the message is built synchronously on the calling thread.
Generally this represents the majority of the call time as sending the
message is generally trivial despite being blocking I/O.

This adds the BuildMessage operation to the background task to minimize
the call time which I would believe to be more correct behavior for the
call.