Use InnerException not GetBaseException
GetBaseException can send you too far down the rabbit hole. For example in a SQL Timeout exception the BaseException is a useless "[Win32Exception: The wait operation timed out]" without a stack trace as it was captured and wrapped by SqlClient. All we want to do here is remove the outermost HttpUnhandledException wrapper