Use ClickOnce data directory as base if application is network deployed
This solves the problem having two client applications running on the same machine using Raygun.
When the network dies all messages for all applications using Raygun stores their reports in the
same directory by default. The reports are sent to Raygun at a later point in time, by the first
application to start up with the Raygun client.

This leads to exception reports being sent to the wrong Raygun account and by isolating the
reports stored on the local hard drive we can at least get the right reports into the right
account at Raygun.io.