Fix nulls in debug info dictionary
There is a possibility that a frame does not have any debugging information attached to it.

This happens when DebugSymbols is set to "None", and the output is in Release mode. The result is DebugInfo but it does not have a signature or other properties.

This addresses that particular scenario, by null checking the debug info as well as the signature before attempting to create the lookup dictionary.